refactored general.h and general.ino. Eliminated duplicate fclamp method in sequencer/misc.h.
