+++ src/filter/adsr.h
@@ -2,7 +2,7 @@
 #ifndef ADSR_H
 #define ADSR_H
 
-#include "../../general.h"
+#include "../general/general.h"
 
 /** Distinct stages that the phase of the envelope can be located in.
 - IDLE   = located at phase location 0, and not currently running
