+++ src/sampler/sampler.cpp
@@ -17,7 +17,6 @@
 
 #include "sampler.h"
 #include "samples.h"
-#include "general.h"
 
 /* You only need to format LittleFS the first time you run a
    test or else use the LittleFS plugin to create a partition
